requirements.txt
torch>=2.3
hydra-core>=1.3.2
pytorch-lightning>=2.2.3
wandb>=0.16.6
rich>=13.7.1
python-dotenv>=1.0.1
scikit-learn>=1.4.2
matplotlib>=3.8.4
